* ext/io/console/console.c (console_set_winsize):   surpress warning: shorten-64-to-32.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@31163 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/io/console/console.c
@@ -414,7 +414,7 @@ console_set_winsize(VALUE io, VALUE size)
 
     GetOpenFile(io, fptr);
     size = rb_Array(size);
-    rb_scan_args(RARRAY_LEN(size), RARRAY_PTR(size), "22",
+    rb_scan_args((int)RARRAY_LEN(size), RARRAY_PTR(size), "22",
                 &row, &col, &xpixel, &ypixel);
 #if defined TIOCSWINSZ
     fd = GetWriteFD(fptr);
